Linear Encoding, Offsets on Sensors and Limiters
Have a rough implementation of linear encoding. It tracks the number of rotations and outputs how many rotations it is from a zero point (in fractions to include partial rotations).

Gave pan an offset so the camera mount's "C" faces approximately to the stepper motor.

Failed to fix bug when hit limit switch

To do:
-Fix zeroing off of initial position for travel
-Fix limit switch bug, or just use a zeroing routine from a library
-Get pan offset to be more accurate
-Integrate with stepper motor code to make testing easier than by hand